Improve install and uninstall commands
